Split out hashutil tests into their own module.
